+++ CHANGELOG.md
@@ -5,8 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
 ### Added
  - Added property to Cisco Codec VideoComponent to determine if the codec supports video mute based on system software version.
+ - Added cameras collection to IConferenceManager
+ - Added features for room-wide camera privacy mute
 
 ## [17.2.1] - 2021-02-18
 ### Changed